fix: renamed citybikes to vehicleRentalUtils
