+++ app/util/vehicleRentalUtils.js
